kopf-requirements.txt
kopf>=1.37.0
kubernetes>=25.0.0
pyyaml>=6.0
